Rename dataset_params.dataset_config to dataset_params.subset (#109)
1. We currently follow HF's naming convention for
`dataset_params.dataset_config`. This allows to specify a subfolder to
load, within the root `dataset_name` folder.

This is commonly used for datasets with multiple subsets (e.g. MMLU). 

In preparation for our own datasets abstraction, this PR renames the
variable to `subset` to better reflect what it actually does instead of
the legacy HF name.

2. Move `device_map` from hardcoded within the model loader to the
config.