Add Dataset base class & API (#151)
PR to add `MapDataset` and `SftDataset` abstractions. Also refactors
Alpaca and ChatQA datasets to use the new abstractions.

Towards OPE-58

**Tasks**
- [x] File structure
- [x] Integration with training loop
- [ ] Iterable datasets for pre-training -> Future PR
- [x] Porting existing SFT datasets code
- [x] Leverage registry
- [x] Support packing
- [ ] Documentation -> Notebook for now, markdown in a future PR
- [x] Tests


**Might change**
- [x] Pandas as default data backend -> Sticking with this for now for
the SFT datasets. There is no visible performance impact (so far), and
it's extremely convenient for data manipulation / debugging.
- [X] Pydantic types -> Keeping for now. This avoids the need to
manually build the data class objects and the strict type verification
was super helpful catching bugs when iterating on the datasets