Replacing `EvaluationConfig`'s `DataParams` with `DatasetSplitParams` (#168)
Initially it seemed that this flexibility (being able to define multiple
dataset splits for evaluation) is useful BUT it may create more
confusion than help (Yes, Matt warned me and I didn't listen). For
instance, when running experiments with LM Harness, users might think
that they are using the **validation** dataset of the benchmark (since
they are adding the `validation` term at
`config.data.validation.datasets[..]` but they are actually NOT! LM
harness is choosing the relevant split for evals. So, I am dropping the
`validation` field to make things cleaner (until we really need this,
which we might not).