Enable Liger for Llama 8B SFT (#439)
With our default config, enabling Liger for 8B SFT has a noticeable
improvement to training speed, which is not the case for other configs.

Also switched to explicitly training for 1 epoch instead of putting the
calculated steps needed. Also adjusted the save steps for the 8B configs
to be more reasonable given the total number of steps and training
speed.