Various updates to Llama 2b configs (#489)
Fixes OPE-307

- Switch to AdamW as optimizer for Llama 2b FSDP for now, given we've
found that adafactor currently doesn't play well with FSDP. We don't
have memory issues with AdamW, and the MFU is unchanged.
- Remove disable_dropout from llama2b config since it's a no-op (tested
that it has no effect)
- Switch Llama 2b config's optim to adafactor since that's what we're
setting everywhere
- Add scale_parameter=False to LEMA Adafactor optim to match HF's
default. I also remove specifying the beta1 param for now, since HF
doesn't seem to do so; if we still want to pass it in, we should
probably rename config.adam_beta1 to be more generic.