Various improvements to Llama eval scripts (#556)
Towards OPE-313, OPE-140

- Add Llama 70b GCP eval config
- Fix bug in arg names passed to eval script
- Deleted legacy Oumi eval config. According to Kostas, there's no
reason anyone should be using that now, so removing it to not confuse
users.
- Clean up unneeded vars in config, and ensure all configs are
consistent
- Add validation that for models that require sharding for eval, that
the job is not invoked by Accelerate.

Tested Llama 8B/70B SFT/LoRA evals on GCP/Polaris. On 1 Polaris node, we
achieve ~240 it/s for 8B and ~7 it/s for 70B. Note that while tqdm shows
60 it/s for 8B, that is per GPU, as we do DDP for 8B eval. For LoRA
evals, it's ~10-20% slower than the above speeds.