[Inference] Bug in generation config stop tokens (#663)
- In `NativeTextInferenceEngine`, we generate a
`transformers.GenerationConfig` from our `generation_params`, where we
pass as "stop" argument the token(s) to terminate the sequence. However,
there is no `stop` argument in the `transformers.GenerationConfig`'s
`__init__`; to the best of my understanding this argument should be a
no-op.
- Since the generation does not have a stop token, it picks up the
default one from Meta. In the model that I am currently investigating
(trained from base), this defaults to `128001`; generation thinks our
model is the base model.
- The right way to determine the stop token is by `stop_strings` or
`eos_token_id`.
- We also need to pass in the tokenizer in `model.generate` kwargs (as
`tokenizer`); this is a requirement if we are adjusting the stop tokens.
- Future work (P1): Offer users the option to pick up the generation
config from the model checkpoint (i.e., `generation_config.json`) and
merge it with the `generation_params` defined in our config.