Allow None batch size in `GenerationParams` (#895)
# Description:

In our tour notebook, we were setting batch size to 0 to supposedly let
LM Harness decide the batch size. This was erroring because we do a
validation check that the batch size is positive. In addition, LM
Harness wasn't actually deciding the optimal batch size.

Per their
[docs](https://github.com/EleutherAI/lm-evaluation-harness/blob/0b994433b86efdf833dae073d8f1ce5537222bd0/docs/interface.md?plain=1#L23),
batch size needs to either be positive or "auto". This PR enables
`batch_size` to be None, in which case our evaluate code sets it to be
`"auto"`. Based on the output logs, I can confirm that this causes LM
Harness to correctly search for the optimal batch size.

I switched the default batch size value to be 1, which IMO makes more
sense as a default than 2. I also think it makes more sense as a default
value than None.

## Related issues

Fixes OPE-726

## Before submitting
- [ ] This PR only changes documentation. (You can ignore the following
checks in that case)
- [x] Did you read the [contributor
guideline](https://github.com/oumi-ai/oumi/blob/main/CONTRIBUTING.md)
Pull Request guidelines?
- [x] Did you link the issue(s) related to this PR in the section above?
- [x] Did you add / update tests where needed?