[evaluations/generative_benchmark] Removing notebook
This notebook was developed before we properly integrated with AlpacaEval. ATM, we do NOT need a notebook to run Alpaca, all we need is running a yaml file (as shown right above). 

Still, if users do want to run AlpacaEval outside Oumi, they can use this notebook, but this is also advertised right above ("If you prefer to use AlpacaEval outside Oumi, we refer you to our example notebook"), so we do not need to re-link it in our resources IMO.