Review changes for the tune module & add custom evaluation metrics
This commit fixes some issues from the tune module and moves the
main training loop to the train.py, adapting it to return the
metrics if the user calls it with additional_tuning_args.

This commit also enables custom evalution metrics for the tuning,
enabling the study to improve custom metrics when tuning.

Also some tests are now corrected to include the new tuning
config.